test: update test for the tablePage
